audio: move test components into the src/samples/audio dir
This commit moves all of the template components into
src/samples/audio directory in order to avoid confusion
during 3rd part IP integration process.

Building samples are set by default only from chrome
configs.
