+++ src/CMakeLists.txt
@@ -6,6 +6,9 @@ add_subdirectory(ipc)
 add_subdirectory(audio)
 add_subdirectory(lib)
 add_subdirectory(math)
+if(CONFIG_SAMPLES)
+	add_subdirectory(samples)
+endif()
 add_local_sources(sof spinlock.c)
 
 if(CONFIG_LIBRARY)
