+++ src/include/sof/samples/audio/detect_test.h
