+++ src/include/sof/samples/audio/smart_amp_test.h
