+++ src/samples/CMakeLists.txt
@@ -1,0 +1,5 @@
+# SPDX-License-Identifier: BSD-3-Clause
+
+if(NOT CONFIG_LIBRARY)
+	add_subdirectory(audio)
+endif()
