+++ src/samples/audio/CMakeLists.txt
@@ -1,0 +1,13 @@
+# SPDX-License-Identifier: BSD-3-Clause
+
+if(CONFIG_SAMPLE_SMART_AMP)
+	add_local_sources(sof
+		smart_amp_test.c
+	)
+endif()
+
+if(CONFIG_SAMPLE_KEYPHRASE)
+	add_local_sources(sof
+		detect_test.c
+	)
+endif()
